requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.5.2
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
pg8000==1.31.2
propcache==0.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
requests==2.32.3
scramp==1.4.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
